[#29,data tables][l]: extract table util functions to separate lib files
